Added BuyActivity.java and SellActivity.java and moved User, Stock, Transaction entities under database/entities.
