+++ app/src/main/java/com/example/datadolphinsandroidapp/BuyActivity.java
@@ -1,0 +1,4 @@
+package com.example.datadolphinsandroidapp;
+
+public class BuyActivity {
+}
