+++ app/src/main/java/com/example/datadolphinsandroidapp/SellActivity.java
@@ -1,0 +1,4 @@
+package com.example.datadolphinsandroidapp;
+
+public class SellActivity {
+}
